requirements.txt
# Core
python-dateutil>=2.1
netaddr>=0.8.0
sqlitedict>=1.6.0
cherrypy>=18.1.0
cherrypy-cors>=1.6
coloredlogs<=10.0
asyncio-throttle==0.1.1

# AWS Provider
botocore>=1.20.21
boto3>=1.9.210
policyuniverse>=1.3.2.0

# GCP Provider
grpcio>=1.18.0
## Cloud Libraries
google-cloud-container>=2.1.0
google-cloud-core>=0.29.1
google-cloud-iam>=0.1.0
google-cloud-logging>=2.2.0
google-cloud-monitoring==1.1.0
google-cloud-resource-manager>=0.28.3
google-cloud-storage>=1.13.2
google-cloud-kms==1.3.0
## API Client Libraries
google-api-python-client>=2.47.0
oauth2client>=4.1.3
## Necessary since API Client Libraries are not thread-safe
httplib2shim>=0.0.3

#for authentication
azure-identity==1.5.0

## for resources

azure-mgmt-resource==15.0.0
azure-mgmt-storage==17.0.0
azure-mgmt-monitor==2.0.0
azure-mgmt-sql==1.0.0
azure-mgmt-security==1.0.0
azure-mgmt-keyvault==8.0.0
azure-mgmt-network==17.1.0
azure-mgmt-redis==12.0.0
azure-mgmt-web==1.0.0
azure-mgmt-compute==18.2.0
azure-mgmt-authorization==3.0.0
azure-mgmt-rdbms==8.0.0

msgraph-core==0.2.2

# Aliyun / Alibaba Cloud Provider
aliyun-python-sdk-core>=2.13.4
aliyun-python-sdk-ram>=3.0.1
aliyun-python-sdk-ocs>=0.0.4
aliyun-python-sdk-sts>=3.0.1
aliyun-python-sdk-actiontrail>=2.0.0
aliyun-python-sdk-vpc>=3.0.5
aliyun-python-sdk-ecs>=4.16.10
aliyun-python-sdk-rds>=2.3.9
aliyun-python-sdk-kms>=2.6.0
oss2>=2.8.0

# Oracle Cloud Infrastructure Provider
oci>=2.2.4

# Kubernetes SDK
kubernetes

# DigitalOcean Cloud Provider
pydo >=0.2.0

# Django
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers
django-prometheus==2.3.1


# Celery
celery==5.4.0
redis==5.0.7

# CrewAI
stripe
python-freshdesk==1.3.9
crewai==0.28.0
crewai-tools
pydantic==2.8.2
PyYAML==6.0.2
structlog==24.1.0
drf-spectacular==0.27.2
